Sửa lỗi và thêm chức năng bên dealer controller

+++ Project/Areas/Admin/Controllers/DealersController.cs
@@ -11,11 +11,11 @@ using Project.Models.DTO;
 using Project.Repositories.Abstract;
 using System.Net;
 using System.Security.Claims;
+using System.Text.RegularExpressions;
 
 namespace Project.Areas.Admin.Controllers
 {
     [Area("Admin")]
-    [Authorize(Roles = "admin")]
     public class DealersController : Controller
     {
         private readonly IUserAuthenticationService _authService;
@@ -29,8 +29,9 @@ namespace Project.Areas.Admin.Controllers
             _userManager = userManager;
 			notyfService = _notyfService;
 		}
-        // GET: DealersController
-        public ActionResult Index()
+		// GET: DealersController
+		[Authorize(Roles = "admin")]
+		public ActionResult Index()
         {
             var model = db.Dealers
                 .Include(d => d.ApplicationUser)
@@ -38,23 +39,64 @@ namespace Project.Areas.Admin.Controllers
             return View(model);
         }
 
-        // GET: DealersController/Details/5
-        public ActionResult Details(int id)
+		[Authorize(Roles = "dealer")]
+		[HttpGet]
+		public IActionResult ChangePassword()
+		{
+			return View();
+		}
+
+		[Authorize(Roles = "dealer")]
+		[HttpPost]
+		public async Task<IActionResult> ChangePassword(ChangePasswordModel model)
+		{
+			if (!ModelState.IsValid)
+			{
+				return View(model);
+
+			}
+
+			if (string.Compare(model.CurrentPassword, model.NewPassword) == 0)
+			{
+				TempData["Error"] = "The new password cannot be the same as the old password.";
+			}
+			else
+			{
+				var result = await _authService.ChangePasswordAsync(model, User.Identity.Name);
+
+
+				if (result.Message.Equals("Some error occcured"))
+				{
+					TempData["Error"] = "Current password is incorrect";
+
+				}
+				else
+				{
+					TempData["success"] = result.Message;
+				}
+			}
+			return RedirectToAction(nameof(ChangePassword));
+
+		}
+
+		// GET: DealersController/Create
+		[Authorize(Roles = "admin")]
+		public ActionResult Create()
         {
             return View();
         }
 
-        // GET: DealersController/Create
-        public ActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: DealersController/Create
-        [HttpPost]
+		// POST: DealersController/Create
+		[Authorize(Roles = "admin")]
+		[HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(RegistrationModel model, string phone, string address)
         {
+            if (!Regex.IsMatch(phone, @"^\d+$"))
+            {
+                TempData["Error"] = "The phone field must be numeric.";
+            }
+
             if (!ModelState.IsValid)
             {
                 return View(model);
@@ -90,20 +132,29 @@ namespace Project.Areas.Admin.Controllers
         }
 
 
-        // GET: DealersController/Edit/5
-        public async Task<IActionResult> Edit(int id)
+		// GET: DealersController/Edit/5
+		[Authorize(Roles = "admin")]
+		public async Task<IActionResult> Edit(int id)
         {
             var deal = db.Dealers.SingleOrDefault(d => d.id.Equals(id));
             return View(deal);
         }
 
         // POST: DealersController/Edit/5
+        [Authorize(Roles = "admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(string dealers_id, string newPassword)
         {
             var deal = db.Dealers.Find(int.Parse(dealers_id));
             var user = await _userManager.FindByIdAsync(deal.user_id);
+
+            if (!IsValidPassword(newPassword))
+            {
+                TempData["Error"] = "Password must contain at least one uppercase letter, one lowercase letter, one number, one special character, and be at least 6 characters long.";
+                return View(deal);
+            }
+
             var token = await _userManager.GeneratePasswordResetTokenAsync(user);
             var result = await _userManager.ResetPasswordAsync(user, token, newPassword);
             try
@@ -113,7 +164,7 @@ namespace Project.Areas.Admin.Controllers
                     // Lưu thay đổi vào cơ sở dữ liệu
                     db.SaveChanges();
                     notyfService.Success("Edit successfully");
-					return RedirectToAction("Index");
+                    return RedirectToAction("Index");
                 }
                 else
                 {
@@ -124,14 +175,22 @@ namespace Project.Areas.Admin.Controllers
             {
                 ModelState.AddModelError(string.Empty, ex.Message);
             }
-            return View();
+            return View(deal);
+        }
+
+        private bool IsValidPassword(string password)
+        {
+            // Kiểm tra password phải có ít nhất 8 ký tự và chứa ít nhất một ký tự hoa, một ký tự thường, một ký tự số và một ký tự đặc biệt
+            var regex = new Regex(@"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^\da-zA-Z]).{6,}$");
+            return regex.IsMatch(password);
         }
 
 
 
 
         // GET: DealersController/Delete/5
-        public async Task<IActionResult> Delete(int id)
+        [Authorize(Roles = "admin")]
+		public async Task<IActionResult> Delete(int id)
         {
             try
             {
@@ -139,6 +198,13 @@ namespace Project.Areas.Admin.Controllers
                 if (dealer == null)
                 {
                     return NotFound();
+                }
+
+                var hasOrders = db.Dealer_Orders.Any(d => d.dealers_id == id);
+                if (hasOrders)
+                {
+                    notyfService.Error("This dealer cannot be deleted because they have placed orders.");
+                    return RedirectToAction("Index");
                 }
 
                 db.Dealers.Remove(dealer);
